# authentication/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/signup": {
      "post": {
        "operationId": "signup",
        "summary": "signup",
        "description": "Default description",
        "query": {},
        "response": {
          "isArray": false,
          "description": ""
        },
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "firstname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "login",
        "summary": "login",
        "description": "Default description",
        "query": {},
        "response": {
          "isArray": false,
          "description": ""
        },
        "parameters": [
          {
            "description": "Username of user",
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Password",
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/refresh": {
      "post": {
        "operationId": "refresh",
        "summary": "refresh",
        "description": "Default description",
        "response": {
          "isArray": false,
          "description": ""
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/me": {
      "get": {
        "operationId": "me",
        "summary": "me",
        "description": "Default description",
        "response": {
          "isArray": false,
          "description": ""
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/swagger.json": {
      "get": {
        "operationId": "AppController_swagger",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/users/profile": {
      "get": {
        "operationId": "getUserProfile",
        "summary": "getUserProfile",
        "description": "Default description",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Authentication Service",
    "description": "Authentication API Swagger",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost/api/authentication"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "firstname",
          "lastname"
        ]
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "description": "Metadata related to the response"
          },
          "data": {
            "type": "object",
            "description": "The actual data of the response"
          },
          "status": {
            "type": "string",
            "description": "The status of the response"
          },
          "message": {
            "type": "string",
            "description": "Message describing the response"
          },
          "errors": {
            "description": "List of errors if any",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username of user"
          },
          "password": {
            "type": "string",
            "description": "Password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      }
    }
  },
  "externalDocs": {
    "description": "More Information",
    "url": "https://www.google.com"
  }
}